requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
dataclasses==0.8
dataclasses-json==0.5.6
discord.py==1.7.3
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
idna-ssl==1.1.0
isodate==0.6.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
multidict==5.2.0
mypy-extensions==0.4.3
oauthlib==3.1.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.6
python-dotenv==0.19.2
python-youtube==0.8.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
typing-extensions==4.0.1
typing-inspect==0.7.1
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
youtube-dl==2021.12.17
